+++ lib/kb_phylogenomics/kb_phylogenomicsClient.py
@@ -78,7 +78,10 @@ class kb_phylogenomics(object):
            "count_category" of String, parameter "heatmap" of type "bool",
            parameter "vertical" of type "bool", parameter "top_hit" of type
            "bool", parameter "e_value" of Double, parameter "log_base" of
-           Double, parameter "show_blanks" of type "bool"
+           Double, parameter "show_blanks" of type "bool", parameter
+           "display_genome_object_name" of type "bool", parameter
+           "skip_missing_genomes" of type "bool", parameter
+           "enforce_genome_version_match" of type "bool"
         :returns: instance of type "view_fxn_profile_Output" -> structure:
            parameter "report_name" of String, parameter "report_ref" of String
         """
@@ -102,7 +105,10 @@ class kb_phylogenomics(object):
            "count_category" of String, parameter "heatmap" of type "bool",
            parameter "vertical" of type "bool", parameter "top_hit" of type
            "bool", parameter "e_value" of Double, parameter "log_base" of
-           Double, parameter "show_blanks" of type "bool"
+           Double, parameter "show_blanks" of type "bool", parameter
+           "display_genome_object_name" of type "bool", parameter
+           "skip_missing_genomes" of type "bool", parameter
+           "enforce_genome_version_match" of type "bool"
         :returns: instance of type "view_fxn_profile_featureSet_Output" ->
            structure: parameter "report_name" of String, parameter
            "report_ref" of String
@@ -128,7 +134,10 @@ class kb_phylogenomics(object):
            "count_category" of String, parameter "heatmap" of type "bool",
            parameter "vertical" of type "bool", parameter "top_hit" of type
            "bool", parameter "e_value" of Double, parameter "log_base" of
-           Double, parameter "show_blanks" of type "bool"
+           Double, parameter "show_blanks" of type "bool", parameter
+           "display_genome_object_name" of type "bool", parameter
+           "skip_missing_genomes" of type "bool", parameter
+           "enforce_genome_version_match" of type "bool"
         :returns: instance of type "view_fxn_profile_phylo_Output" ->
            structure: parameter "report_name" of String, parameter
            "report_ref" of String
@@ -221,7 +230,9 @@ class kb_phylogenomics(object):
            "workspace_name" (** Common types), parameter
            "input_pangenome_ref" of type "data_obj_ref", parameter
            "input_speciesTree_ref" of type "data_obj_ref", parameter
-           "save_featuresets" of type "bool"
+           "save_featuresets" of type "bool", parameter
+           "skip_missing_genomes" of type "bool", parameter
+           "enforce_genome_version_match" of type "bool"
         :returns: instance of type "view_pan_phylo_Output" -> structure:
            parameter "report_name" of String, parameter "report_ref" of String
         """
@@ -270,7 +281,10 @@ class kb_phylogenomics(object):
            "count_category" of String, parameter "heatmap" of type "bool",
            parameter "vertical" of type "bool", parameter "top_hit" of type
            "bool", parameter "e_value" of Double, parameter "log_base" of
-           Double, parameter "show_blanks" of type "bool"
+           Double, parameter "show_blanks" of type "bool", parameter
+           "display_genome_object_name" of type "bool", parameter
+           "skip_missing_genomes" of type "bool", parameter
+           "enforce_genome_version_match" of type "bool"
         :returns: instance of type "get_configure_categories_Output" ->
            structure: parameter "cats" of list of String, parameter
            "cat2name" of type "Cat2Name" (category to name) -> structure: